requirements.txt
BioPython
numpy
tqdm
time
sklearn
csv
matplotlib
pickle
